stix_to_query->results_translator
