+++ adapter-guide/develop-stix-adapter.md
@@ -74,7 +74,7 @@ To develop a STIX-shifter connector for a data source:
 | data_mapping.py         | This file uses the mappings that are defined in the from_stix_map.json file to map the STIX objects and their properties to field names from the data source.                                                                                                                                              |
 | dummy_translator.py     | This file contains the Translator class. It inherits the BaseTranslator abstract base class and is the interface to the rest of the translation logic.                                                                                                                                                     |
 | query_constructor.py    | This file contains the QueryStringPatternTranslator class, which translates the ANTLR parsing of the STIX pattern to the native data source query.                                                                                                                                                         |
-| stix_to_query.py        | This file contains the StixToQuery class, which inherits the BaseQueryTranslator class. <br><br>StixToQuery calls out to the ANTLR parser, which returns a parsing of the STIX pattern. The parsing is then passed onto the query_constructor.py where it is translated into the native data source query. |
+| query_translator.py        | This file contains the QueryTranslator class, which inherits the BaseQueryTranslator class. <br><br>QueryTranslator calls out to the ANTLR parser, which returns a parsing of the STIX pattern. The parsing is then passed onto the query_constructor.py where it is translated into the native data source query. |
 | transformers.py         | This file is used to transform data formats as required by STIX and the native data source query language.                                                                                                                                                                                                 |
 | MANIFEST.in             | This file is used by Python when packaging the library.                                                                                                                                                                                                                                                    |
