+++ stix_shifter_modules/async_dummy/stix_translation/query_translator.py
@@ -1,12 +1,12 @@
 import logging
 
-from stix_shifter_utils.modules.base.stix_translation.base_translator import BaseQueryTranslator
+from stix_shifter_utils.modules.base.stix_translation.base_query_translator import BaseQueryTranslator
 from . import query_constructor
 
 logger = logging.getLogger(__name__)
 
 
-class StixToQuery(BaseQueryTranslator):
+class QueryTranslator(BaseQueryTranslator):
 
     def transform_query(self, data, antlr_parsing_object, data_model_mapper, options, mapping=None):
         """
@@ -21,7 +21,7 @@ class StixToQuery(BaseQueryTranslator):
         :rtype: str
         """
 
-        logger.info("Converting STIX2 Pattern to ariel")
+        logger.info("Converting STIX2 Pattern to data source query")
 
         query_string = query_constructor.translate_pattern(
             antlr_parsing_object, data_model_mapper, options)
