+++ stix_shifter_modules/bigfix/stix_translation/query_translator.py
@@ -6,7 +6,7 @@ from . import query_constructor
 logger = logging.getLogger(__name__)
 
 
-class StixToQuery(BaseQueryTranslator):
+class QueryTranslator(BaseQueryTranslator):
 
     def transform_query(self, data, antlr_parsing_object, data_model_mapper, options, mapping=None):
         """
