+++ stix_shifter_modules/carbonblack/stix_translation/query_translator.py
@@ -1,12 +1,12 @@
 import logging
 
-from stix_shifter_utils.modules.base.stix_translation.base_translator import BaseQueryTranslator
+from stix_shifter_utils.modules.base.stix_translation.base_query_translator import BaseQueryTranslator
 from . import query_constructor
 
-LOGGER = logging.getLogger(__name__)
+logger = logging.getLogger(__name__)
 
 
-class StixToQuery(BaseQueryTranslator):
+class QueryTranslator(BaseQueryTranslator):
 
     def transform_query(self, data, antlr_parsing_object, data_model_mapper, options, mapping=None):
         """
@@ -15,14 +15,13 @@ class StixToQuery(BaseQueryTranslator):
         :type antlr_parsing_object: object
         :data_model_mapper: Mapping object for the data source
         :type data_model_mapper: object
-        :param mapping: The mapping file path to use as instructions on how to transform
-        the given STIX query into another format. This should default to something if one isn't passed in
+        :param mapping: The mapping file path to use as instructions on how to transform the given STIX query into another format. This should default to something if one isn't passed in
         :type mapping: str (filepath)
         :return: transformed query string
         :rtype: str
         """
 
-        LOGGER.info("Converting STIX2 Pattern to data source query")
+        logger.info("Converting STIX2 Pattern to cbquery")
 
         query_string = query_constructor.translate_pattern(
             antlr_parsing_object, data_model_mapper, options)
