+++ stix_shifter_modules/elastic_ecs/stix_translation/query_translator.py
@@ -3,14 +3,15 @@ import logging
 from stix_shifter_utils.modules.base.stix_translation.base_query_translator import BaseQueryTranslator
 from . import query_constructor
 
+
 logger = logging.getLogger(__name__)
 
 
-class StixToQuery(BaseQueryTranslator):
+class QueryTranslator(BaseQueryTranslator):
 
     def transform_query(self, data, antlr_parsing_object, data_model_mapper, options, mapping=None):
         """
-        Transforms STIX pattern into a different query format. Based on a mapping file
+        Transforms STIX query into a different query format. Based on a mapping file
         :param antlr_parsing_object: Antlr parsing objects for the STIX pattern
         :type antlr_parsing_object: object
         :data_model_mapper: Mapping object for the data source
@@ -21,7 +22,7 @@ class StixToQuery(BaseQueryTranslator):
         :rtype: str
         """
 
-        logger.info("Converting STIX2 Pattern to cbquery")
+        logger.info("Converting STIX2 Pattern to data source query")
 
         query_string = query_constructor.translate_pattern(
             antlr_parsing_object, data_model_mapper, options)
