+++ stix_shifter_modules/guardium/stix_translation/query_translator.py
@@ -5,9 +5,8 @@ from . import query_constructor
 #logger = #logging.getLogger(__name__)
 
 
-class StixToQuery(BaseQueryTranslator):
+class QueryTranslator(BaseQueryTranslator):
 
-    #def transform_query(self, data, options, mapping=None):
     def transform_query(self, data, antlr_parsing_object, data_model_mapper, options, mapping=None):
         """
         Transforms STIX query into data source query format. Based on a mapping file
