--- stix_shifter_modules/synchronous_dummy/stix_translation/dummy_translator.py
@@ -1,15 +1,0 @@
-from stix_shifter_utils.modules.base.stix_translation.base_translator import BaseTranslator
-from .stix_to_query import StixToQuery
-from stix_shifter_utils.stix_translation.src.json_to_stix.json_to_stix import JSONToStix
-from os import path
-
-
-class Translator(BaseTranslator):
-
-    def __init__(self, dialect=None):
-        basepath = path.dirname(__file__)
-        filepath = path.abspath(
-            path.join(basepath, "json", "to_stix_map.json"))
-        self.mapping_filepath = filepath
-        self.results_translator = JSONToStix(filepath)
-        self.query_translator = StixToQuery()
