+++ tests/stix_translation/test_azure_sentinel_stix_to_query.py
@@ -13,7 +13,7 @@ def _remove_timestamp_from_query(queries):
         return re.sub(pattern, "", queries)
 
 
-class TestStixToQuery(unittest.TestCase):
+class TestQueryTranslator(unittest.TestCase):
     """
     class to perform unit test case azure_sentinel translate query
     """
