+++ tests/stix_translation/test_security_advisor_stix_to_query.py
@@ -14,7 +14,7 @@ def _test_query_assertions(query, queries):
     assert query['queries'] == queries
 
 
-class TestStixToQuery(unittest.TestCase):
+class TestQueryTranslator(unittest.TestCase):
 
     def test_simple_or_query(self):
         stix_pattern = "[url:value = 'http://5.188.86.29:7000' OR url:value = 'http://5.45.69.149:7000']"
